Separate Header into its own component and add 10 quotes with authors for dynamic display
