+++ src/App.jsx
@@ -1,46 +1,11 @@
 import "./App.css";
-import { useState } from "react";
-import ModalComponent from "./components/Modal/Modal";
+import Header from "./features/header/Header";
+import ModalComponent from "./components/ModalComponent/ModalComponent";
 
 const App = () => {
-  const [isModalOpen, setIsModalOpen] = useState(false);
-
-  const openModal = () => setIsModalOpen(true);
-  const closeModal = () => setIsModalOpen(false);
-
   return (
     <div className="page-wrap">
-      {/* header */}
-      <header className="header">
-        <div className="wrap">
-          <span className="btn-icon" onClick={openModal}>
-            <img
-              className="icon icon-plus js-modal-init"
-              src="icons/icon-plus.svg"
-              alt="Add New Item"
-            />
-          </span>
-          <div className="header-blockquote">
-            <h1 className="header-quote">
-              Happiness is good health and a bad memory
-            </h1>
-            <div className="header-cite">- Ingrid Bergman</div>
-          </div>
-        </div>
-        <div className="header-inner">
-          <div className="wrap">
-            <img className="logo" src="images/vegait-logo.svg" alt="VegaIT" />
-            <div className="date-wrap">
-              <img
-                className="icon"
-                src="icons/icon-calendar.svg"
-                alt="Calendar"
-              />
-              <time>02 / 08 / 2019</time>
-            </div>
-          </div>
-        </div>
-      </header>
+      <Header />
       <main className="main">
         <div className="wrap">
           {/* item */}
@@ -84,7 +49,6 @@ const App = () => {
       </main>
 
       {/* Modal Component */}
-      <ModalComponent isModalOpen={isModalOpen} closeModal={closeModal} />
 
       {/* footer */}
       <footer className="footer">
